Export excel

+++ app/Services/LiquidationService.php
@@ -8,6 +8,7 @@ use App\Repositories\LiquidationRepository;
 use App\Repositories\DetailImportStoreRepository;
 use App\Repositories\StoreFacultyRepository;
 use App\Repositories\StoreRoomRepository;
+use App\Repositories\RequestRepository;
 
 class LiquidationService extends BaseService
 {
@@ -23,6 +24,8 @@ class LiquidationService extends BaseService
     private $storeFacultyRepo;
     
     private $storeRoomRepo;
+    
+    private $requestRepo;
 
     /**
      * Constructor of liquidation service
@@ -31,17 +34,20 @@ class LiquidationService extends BaseService
      * @param DetailImportStoreRepository $detailIStoreRepo []
      * @param StoreFacultyRepository      $storeFacultyRepo []
      * @param StoreRoomRepository         $storeRoomRepo    []
+     * @param RequestRepository           $requestRepo      []
      */
     public function __construct(
         LiquidationRepository $liquidationRepo,
         DetailImportStoreRepository $detailIStoreRepo,
         StoreFacultyRepository $storeFacultyRepo,
-        StoreRoomRepository $storeRoomRepo
+        StoreRoomRepository $storeRoomRepo,
+        RequestRepository $requestRepo
     ) {
         $this->liquidationRepo = $liquidationRepo;
         $this->detailIStoreRepo = $detailIStoreRepo;
         $this->storeFacultyRepo = $storeFacultyRepo;
         $this->storeRoomRepo = $storeRoomRepo;
+        $this->requestRepo = $requestRepo;
     }
     
     /**
@@ -82,15 +88,18 @@ class LiquidationService extends BaseService
      *
      * @return void
      */
-    public function removeToLiquordation($detailId)
+    public function removeToLiquidation($detailId)
     {
         $detail = $this->detailIStoreRepo->find($detailId);
         $datas = [
+            'date_liquidation' => Carbon::now()->format(config('define.date_format')),
             'quantity' => $detail->quantity,
-            'detail_import_store_id' => $detail->id,
-            'date_liquidation' => Carbon::now()->format(config('define.date_format'))
+            'store_liquidation_id' => $detail->id,
+            'store_type' => config('constant.type_school'),
         ];
-        $detail->delete();
+        if ($detail->quantity_start == $detail->quantity) {
+            $detail->delete();
+        }
         $this->liquidationRepo->create($datas);
     }
 
@@ -129,6 +138,25 @@ class LiquidationService extends BaseService
             'date_liquidation' => Carbon::now()->format(config('define.date_format'))
         ];
         $storeRoom->delete();
+        $this->liquidationRepo->create($datas);
+    }
+    
+    /**
+     * Create liquidation by request
+     *
+     * @param type $requestId []
+     *
+     * @return void
+     */
+    public function createLiquidationByRequest($requestId)
+    {
+        $request = $this->requestRepo->find($requestId);
+        $datas = [
+            'date_liquidation' => Carbon::now()->format(config('define.date_format')),
+            'quantity' => $request->quantity,
+            'store_liquidation_id' => $request->store_type_id,
+            'store_type' => $request->type,
+        ];
         $this->liquidationRepo->create($datas);
     }
 }
